# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Syrian Car Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get user favorites
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"

    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add car to favorites
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteBody"
      responses:
        "201":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"
  /favorites/{carId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove car from favorites
      security:
        - bearerAuth: []
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed from favorites
  /chats:
    get:
      operationId: getConversations
      tags: [chat]
      summary: Get user conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /chats/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chats/start:
    post:
      operationId: startConversation
      tags: [chat]
      summary: Start a conversation with a seller about a car
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationBody"
      responses:
        "201":
          description: Conversation created or existing returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/user/{userId}:
    get:
      operationId: getUserReviews
      tags: [reviews]
      summary: Get reviews for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        profilePhoto:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - isVerified
        - isPremium
        - createdAt
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        profilePhoto:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        listingCount:
          type: integer
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - isVerified
        - isPremium
        - listingCount
        - reviewCount
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        profilePhoto:
          type: string
        province:
          type: string
        city:
          type: string
        phone:
          type: string
...
    Car:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerPhoto:
          type: ["string", "null"]
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        province:
          type: string
        city:
          type: string
        saleType:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isHighlighted:
          type: boolean
        isActive:
          type: boolean
        primaryImage:
          type: ["string", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - brand
        - model
        - year
        - price
        - mileage
        - fuelType
        - transmission
        - province
        - city
        - saleType
        - category
        - isFeatured
        - isHighlighted
        - isActive
        - viewCount
        - createdAt
...
    Favorite:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        carId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - carId
        - createdAt
    AddFavoriteBody:
      type: object
      properties:
        carId:
          type: integer
      required:
        - carId
    Conversation:
      type: object
      properties:
        id:
          type: integer
        carId:
          type: integer
        carBrand:
          type: string
        carModel:
          type: string
        carYear:
          type: integer
        carImage:
          type: ["string", "null"]
        otherUserId:
          type: integer
        otherUserName:
          type: string
        otherUserPhoto:
          type: ["string", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - carId
        - carBrand
        - carModel
        - carYear
        - otherUserId
        - otherUserName
        - unreadCount
        - createdAt
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        messageType:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - senderId
        - senderName
        - content
        - messageType
        - isRead
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        messageType:
          type: string
      required:
        - content
    StartConversationBody:
      type: object
      properties:
        sellerId:
          type: integer
        carId:
          type: integer
      required:
        - sellerId
        - carId
    Review:
      type: object
      properties:
        id:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        reviewerPhoto:
          type: ["string", "null"]
        reviewedUserId:
          type: integer
        rating:
          type: number
        comment:
          type: ["string", "null"]
        responseSpeed:
          type: ["integer", "null"]
        trustLevel:
          type: ["integer", "null"]
        listingAccuracy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reviewerId
        - reviewerName
        - reviewedUserId
        - rating
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        reviewedUserId:
          type: integer
        rating:
          type: number
        comment:
          type: string
        responseSpeed:
          type: integer
        trustLevel:
          type: integer
        listingAccuracy:
          type: integer
      required:
        - reviewedUserId
        - rating
...