# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Syrian Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface Favorite {
  id: number;
  userId: number;
  carId: number;
  createdAt: Date;
}